added 2 decimal places

+++ main.py
@@ -156,7 +156,7 @@ class Controller(udi_interface.Node):
 
     def commandAO1(self, command):
 #        mynode = self.poly.getNode('controller')
-        mymessage = int(command.get("value"))
+        mymessage = float(command.get("value"))
         data = {}
         data['AO1'] = mymessage
         json_data = json.dumps(data)
@@ -167,7 +167,7 @@ class Controller(udi_interface.Node):
             )
 
     def commandAO2(self, command):
-        mymessage = int(command.get("value"))
+        mymessage = float(command.get("value"))
         data = {}
         data['AO2'] = mymessage
         json_data = json.dumps(data)
@@ -178,7 +178,7 @@ class Controller(udi_interface.Node):
             )
 
     def commandAO3(self, command):
-        mymessage = int(command.get("value"))
+        mymessage = float(command.get("value"))
         data = {}
         data['AO3'] = mymessage
         json_data = json.dumps(data)
@@ -189,7 +189,7 @@ class Controller(udi_interface.Node):
             )
 
     def commandAO4(self, command):
-        mymessage = int(command.get("value"))
+        mymessage = float(command.get("value"))
         data = {}
         data['AO4'] = mymessage
         json_data = json.dumps(data)
@@ -200,7 +200,7 @@ class Controller(udi_interface.Node):
             )
 
     def commandAO5(self, command):
-        mymessage = int(command.get("value"))
+        mymessage = float(command.get("value"))
         data = {}
         data['AO5'] = mymessage
         json_data = json.dumps(data)
